rename abstract-bit-json to abstract-bit-config
