+++ src/consumer/bit-json/abstract-bit-config.js
@@ -256,7 +256,7 @@ export default class AbstractBitConfig {
   static async removeIfExist(bitPath: string): Promise<boolean> {
     const dirToRemove = this.composePath(bitPath);
     if (fs.exists(dirToRemove)) {
-      logger.info(`abstract-bit-json, deleting ${dirToRemove}`);
+      logger.info(`abstract-bit-config, deleting ${dirToRemove}`);
       return fs.remove(dirToRemove);
     }
     return false;
